refactor: updateEventDataObjects
