+++ schedule/src/utils/addCurrenttimeToEvents.js
