+++ schedule/src/utils/updateEventDataObjects.js
@@ -1,0 +1,11 @@
+import addCurrentTimeToEvents from './addCurrenttimeToEvents';
+import createDeadlineEvents from './createDeadlineEvents';
+
+const updateEventDataObjects = (data, timezone) => {
+    return addCurrentTimeToEvents(
+        createDeadlineEvents(data),
+        timezone,
+    )
+};
+
+export default updateEventDataObjects;
